clearing branch
